requirements.txt
astroid==1.6.2
autopep8==1.3.5
click==6.7
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.7.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
pep8==1.7.1
pycodestyle==2.3.1
pylint==1.8.3
pytz==2017.3
six==1.11.0
wrapt==1.10.11
